Allow to disable port on services.
When using service as load balancer, if we setup ldap to be on secure
port only we don't want to expose the insecure port.
Currently there is no way to do that.

With this modification, users can enable or disable exposition of
insecure or secure port on the service.

Default values keep the current behaviour.

Signed-off-by: Guillaume Perrin <guillaume28.perrin@gmail.com>